split formdommanager and worked in clearform after all save actions
